+++ feincms3/__init__.py
@@ -1,3 +1,3 @@
 # VERSION = (0, 41, 0)
-VERSION = (0, "90b3")
+VERSION = (0, "90b4")
 __version__ = ".".join(map(str, VERSION))